move autoencoders to autoencoders_modulev3
