+++ max/python/max/pipelines/architectures/autoencoders_modulev3/__init__.py
