+++ max/python/max/pipelines/architectures/autoencoders_modulev3/autoencoder_kl.py
