+++ max/python/max/pipelines/architectures/autoencoders_modulev3/autoencoder_kl_qwen_image.py
