+++ max/python/max/pipelines/architectures/autoencoders_modulev3/layers/downsampling.py
