+++ max/python/max/pipelines/architectures/flux1_modulev3/pipeline_flux.py
@@ -32,7 +32,7 @@ from max.pipelines.lib.interfaces import (
 )
 from max.pipelines.lib.interfaces.diffusion_pipeline import max_compile
 
-from ..autoencoders import AutoencoderKLModel
+from ..autoencoders_modulev3 import AutoencoderKLModel
 from ..clip import ClipModel
 from ..t5 import T5Model
 from .model import Flux1TransformerModel
