+++ max/python/max/pipelines/architectures/flux2/pipeline_flux2.py
@@ -27,12 +27,12 @@ from max.pipelines.lib.interfaces.diffusion_pipeline import max_compile
 from max.pipelines.lib.utils import BoundedCache
 from max.profiler import Tracer, traced
 
-from ..autoencoders import AutoencoderKLFlux2Model
+from ..autoencoders_modulev3 import AutoencoderKLFlux2Model
 from ..mistral3.text_encoder import Mistral3TextEncoderModel
 from .model import Flux2TransformerModel
 
 if TYPE_CHECKING:
-    from ..autoencoders.vae import DiagonalGaussianDistribution
+    from ..autoencoders_modulev3.vae import DiagonalGaussianDistribution
 
 
 @dataclass(kw_only=True)
