+++ max/python/max/pipelines/architectures/flux2_modulev3/pipeline_flux2.py
@@ -31,10 +31,9 @@ from max.pipelines.lib.interfaces.diffusion_pipeline import max_compile
 from max.pipelines.lib.utils import BoundedCache
 from max.profiler import Tracer, traced
 
-from ..autoencoders import AutoencoderKLFlux2Model
+from ..autoencoders_modulev3 import AutoencoderKLFlux2Model
 from ..mistral3.text_encoder import Mistral3TextEncoderModel
 from .model import Flux2TransformerModel
-
 
 @dataclass(kw_only=True)
 class Flux2ModelInputs:
